+++ CHANGELOG.md
@@ -8,7 +8,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
--   added support for `react-flow` v12 for `NodeContent` component, added v12-related components: `EdgeDefaultV12`, `NodeDefaultV12`, `EdgeDefs`
+-   added support for React Flow v12
+    -   `<NodeContent />` can used with `flowVersion="v12"`
+    -   more v12-only components: `EdgeDefaultV12`, `NodeDefaultV12`, `EdgeDefs`
+        -   they may be removed in future version when v12 elements are available direcly via `<EdgeDefault />` and `<NodeDefault />`
+
+### Deprecated
+
+-   `<EdgeDefaultV12 />` and `<NodeDefaultV12 />` will be removed when React Flow v12 is supported directly by `<EdgeDefault />` and `<NodeDefault />`
+-   `flowVersion`: `legacy` and `next` will be removed/replaced by `v##` values
 
 ## [24.2.0] - 2025-06-04